fix: expose package-private members after watchlist package reorganisation
