+++ src/main/java/edu/ntnu/idatt2003/millions/view/dashboard/watchlist/dialog/WatchlistNoteDialog.java
@@ -5,7 +5,6 @@ import edu.ntnu.idatt2003.millions.model.stock.Stock;
 import edu.ntnu.idatt2003.millions.util.LanguageManager;
 import edu.ntnu.idatt2003.millions.view.component.Modal;
 import edu.ntnu.idatt2003.millions.view.component.ModalActions;
-import edu.ntnu.idatt2003.millions.view.dashboard.watchlist.DiscardChangesDialog;
 import edu.ntnu.idatt2003.millions.view.dashboard.watchlist.card.StockInfoCard;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextArea;
